Create CASS
An oversampling method for imbalanced data: Conservative Adaptive Synthetic Sampling(CASS).In addition, we have open-sourced a code implementation of SFROR, an evaluation metric that can be used to evaluate the quality of synthetic artificial samples. Finally, we have publicly released the TTD2022 dataset to scholars worldwide, and it is publicly available to interested researchers at http://mad-net.cn:8765.